requirements.txt
pillow
tabulate
pandas
numpy==1.19.5
datasets==1.7.0
scikit-learn
tqdm
transformers
accelerate
torchinfo==1.5.3
boto3
openpyxl
emoji
nltk
tensorboard
seaborn
#fasttext==0.9.1
xgboost
wandb
gensim
streamlit
azure-ai-textanalytics==5.2.0